+++ lib/src/obsoleted/rules/double_literal_format_rule.dart
@@ -31,6 +31,7 @@ class DoubleLiteralFormatRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override