+++ lib/src/obsoleted/rules/no_boolean_literal_compare_rule.dart
@@ -35,6 +35,7 @@ class NoBooleanLiteralCompareRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override