+++ lib/src/obsoleted/rules/no_equal_then_else.dart
@@ -21,6 +21,7 @@ class NoEqualThenElse extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override