+++ lib/src/obsoleted/rules/no_object_declaration.dart
@@ -20,6 +20,7 @@ class NoObjectDeclarationRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override