+++ lib/src/obsoleted/rules/obsolete_rule.dart
@@ -5,13 +5,15 @@ import '../../rules/rule.dart';
 abstract class ObsoleteRule extends Rule {
   final Uri documentationUrl;
 
-  ObsoleteRule({
+  const ObsoleteRule({
     required String id,
     required this.documentationUrl,
     required Severity severity,
+    required Iterable<String> excludes,
   }) : super(
           id: id,
           documentation: const RuleDocumentation(name: '', brief: ''),
           severity: severity,
+          excludes: excludes,
         );
 }