+++ lib/src/obsoleted/rules/prefer_conditional_expressions.dart
@@ -24,6 +24,7 @@ class PreferConditionalExpressions extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override