+++ lib/src/obsoleted/rules/provide_correct_intl_args.dart
@@ -24,6 +24,7 @@ class ProvideCorrectIntlArgsRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override