+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* Add static code diagnostic `avoid-returning-widgets`.
+- Add excludes for a separate rule.
+- Add static code diagnostic `avoid-returning-widgets`.
 
 ## 3.0.0